Update events.lua
This is Getting the Count of Players Using the GlobalState['Count:Players'] this is really helpful to add to scripts and pull the data without the need to loop if using StateBagChangeHandler as well as needing to go to Server to get the player count. Been Using this for a while and is very useful. I also use this same concept for Job OnDuty Count as well for example GlobalState['Count:Police']. Will do a PR on that as well in future.